requirements.txt
pandas == 1.2.4
seaborn == 0.11.1
matplotlib.pyplot == 3.3.4
streamlit == 1.1.0